+++ .gitignore
@@ -1,3 +1,6 @@
+wandb/
+*.iml
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
