+++ requirements.txt
@@ -2,3 +2,4 @@ CairoSVG==2.4.2
 flatland-rl==2.1.10
 pycairo==1.19.1
 pyhumps==1.3.1
+gputil==1.4.0
